Generify separated imploder and tokenizer
