+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/imploder/TreeTokenizer.java
@@ -1,0 +1,82 @@
+package org.spoofax.jsglr2.imploder;
+
+import org.metaborg.parsetable.IProduction;
+import org.spoofax.jsglr.client.imploder.IToken;
+import org.spoofax.jsglr2.parser.Position;
+import org.spoofax.jsglr2.tokens.Tokens;
+
+public abstract class TreeTokenizer<Tree> implements ITokenizer<Tree> {
+    class SubTree {
+        Tree tree;
+        IToken leftToken;
+        IToken rightToken;
+        Position endPosition;
+
+        SubTree(TreeImploder.SubTree<Tree> tree, IToken leftToken, IToken rightToken, Position endPosition) {
+            this.tree = tree.tree;
+            this.leftToken = leftToken;
+            this.rightToken = rightToken;
+            this.endPosition = endPosition;
+            storeTokensInTree(tree.tree, tree.production, leftToken, rightToken);
+        }
+
+    }
+
+    @Override public Tree tokenize(Tokens tokens, TreeImploder.SubTree<Tree> tree) {
+        tokens.makeStartToken();
+        tokenTreeBinding(tokens.startToken(), tree.tree);
+
+        SubTree res = tokenizeInternal(tokens, tree, new Position(0, 1, 1));
+
+        tokens.makeEndToken(new Position(res.endPosition.offset, res.endPosition.line, res.endPosition.column));
+        tokenTreeBinding(tokens.endToken(), res.tree);
+
+        return res.tree;
+    }
+
+    private SubTree tokenizeInternal(Tokens tokens, TreeImploder.SubTree<Tree> tree, Position startPosition) {
+        if(tree.children.size() == 0) {
+            if(tree.width > 0) {
+                Position endPosition = startPosition.step(tokens.getInput(), tree.width);
+                IToken token = tokens.makeToken(startPosition, endPosition, tree.production);
+                tokenTreeBinding(token, tree.tree);
+                return new SubTree(tree, token, token, endPosition);
+            }
+            return new SubTree(tree, null, null, startPosition);
+        } else {
+            IToken leftToken = null;
+            IToken rightToken = null;
+            Position pivotPosition = startPosition;
+            for(TreeImploder.SubTree<Tree> child : tree.children) {
+                SubTree subTree = tokenizeInternal(tokens, child, pivotPosition);
+
+                // If child tree had tokens that were not yet bound, bind them
+                if(child.tree == null) {
+                    if(subTree.leftToken != null)
+                        tokenTreeBinding(subTree.leftToken, tree.tree);
+
+                    if(subTree.rightToken != null)
+                        tokenTreeBinding(subTree.rightToken, tree.tree);
+                }
+
+                // The left-most token of this tree is the first non-null leftToken of a subTree
+                if(leftToken == null)
+                    leftToken = subTree.leftToken;
+
+                // The right-most token of this tree is the last non-null rightToken of a subTree
+                if(subTree.rightToken != null)
+                    rightToken = subTree.rightToken;
+
+                // If tree production == null, that means it's an "amb" node; in that case, position is not advanced
+                if(tree.production != null)
+                    pivotPosition = subTree.endPosition;
+            }
+            return new SubTree(tree, leftToken, rightToken, pivotPosition);
+        }
+    }
+
+    protected abstract void storeTokensInTree(Tree tree, IProduction production, IToken leftToken, IToken rightToken);
+
+    protected abstract void tokenTreeBinding(IToken token, Tree term);
+
+}
