+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/imploder/treefactory/ITreeFactory.java
@@ -1,4 +1,4 @@
-package org.spoofax.jsglr2.imploder;
+package org.spoofax.jsglr2.imploder.treefactory;
 
 public interface ITreeFactory<T> {
